Move module environment getter to its own home
